Add files via upload
This code looks at Z-score for the dHPC stimulation with NAc photometry recording experiments in home cage. The dHPC is stimulated every 30 seconds for a duration of 30 ms for a total of 30 minutes. This code requires the users to manually score the behavior using OpenScope offline by re-watching the video.